Patched IPython so that it uses '1' variants in
    sip.setapi('QString', 1)
    sip.setapi('QVariant', 1)

instead of '2' variant.

This patch is in file /ipython-0.13/IPython/external/qt.py

It is needed for the IPythonQt-based plotting to need in the GUI.

So far everything seems to run well with this patch, since we do not use pyside!

If problems occur, another solution has to be found.




git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/trunk@11209 a23ab9f9-1ea2-4934-960d-990259f9ea8c