propagate setting descriptor to children
this fixes a major bug. When a descriptor of a box was set,
it was not propagated to the children of this very box.
Only when a value was set to the box, suddenly the descriptor
was applied.

Now we properly set the descriptor once it's set everywhere.
This may break many things, but fixed #777.