This commit hopefully fixes the issues mentioned in bug #837
The tracking of heartbeats had to be re-written in larger parts and is now tightly associated to establishing and releasing connections.
Another related problem seemed to be the asynchronous (via broker) call to a just dying server in the deviceServer main application. This call got replaced by a synchronous one and the sleep statement got removed.
Some adaptations on related classes (DeviceClient and DeviceServer) had to be performed.

git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/branches/1.2@14296 a23ab9f9-1ea2-4934-960d-990259f9ea8c