guess device class from project
if a device in a project is offline, we guess that it probably will
eventually be instantiated, and already get a class schema for it,
if we have one.

If different projects define the device differently, this might lead
to weird effects, but well, that doesn't make much sense anyways.

This may be related to
XFEL #4657