removed a race while gathering instances
when the device client gathered available instances, it first
pinged everyone, waited for replies, and then connected the
slot slotInstanceNew, which registers devices which come later.

If now a device comes alive while it is waiting for ping responses,
this device won't be in the final list, as it didn't hear the first
ping, but also won't be listed later because the slotInstanceNew
already passed.