temporary fix running and shutting down macros
macros could not be started and shut down properly. This was
mostly because they seem to sometimes show up as devices in the
GUI server, not as macros. Dunno why.

This makes sure the GUI doesn't get confused by that. Let's hope that
works. Real fix would be to fix the GUI server to treat macros as such.