NetworkAppender now sends a Hash and not a string.
This avoids the need of delimiters whose appearence in the message should be
(but was not) checked and treated.
The Hash contains a vector<Hash>, i.e. a Hash for each message. Its keys are
"timestamp", "type", "category" and "message".
Also the format for the underlying krb_log4cpp logging classes is now
hardcoded here (and not in the device server).