make cli.py importable
cli.py is actually not a module, but a script. But like with
every script, it should still be importable, with an if statement
that is only run if the module is run as a script.

The if statement is non-standard, as the script is to be run from
within ipython, not on the command line.