faster trendline
the current trendline is a big performance problem, as it does too much
and especially requests history too often.

This patch decimates the points already in the GUI and does not depend
on the logger to do it. It also only updates if the GUI has nothing
else to do.