Dummy commit to give info about previous one: Merge local and global slots.
In C++ and Python API1 there was a distinction between local and global slots:
Global slots could only be addressed with "*" and local slots only with a
specific instance id.

Now there are just KARABO_SLOTs - and they can be addressed in both ways. The
only difference is that in case of a request (not a requestNoWait!), the reply
of a globally called slot is ignored.

Now slotPing does not have to be broadcasted if used to find about the
existence of an instanceId. But that required a little extra hack to avoid
that a SignalSlotable that comes up answers (and thus blocks) itself.

While adjusting API2, the timeout to wait at startup time for a potential reply
of another instance with the same id has been adjusted to 1 s as used in C++.

On the fly, the DeviceServer has been cleaned up a bit.


git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/trunk@19041 a23ab9f9-1ea2-4934-960d-990259f9ea8c