Import project-related bits directly from karabo.api2.project.
This is actually solving the fact that karabo.api_2 provides a Monitor which
is a function decorator instead of a class for monitor objects. Really though,
the GUI should be free to import directly from the karabo package instead of
indirectly via karabo.api_2. These two packages live in the same source code
repository and should always be kept in sync with each other. Whether we
successfully keep them in sync is another matter entirely.

Some unused imports in guiproject.py were also removed by this commit.