introduce a main method for devices
Until now, a device was several tasks running in parallel, all doing
their stuff. This changes that there is one task which defines the
runtime of the device. This task is the one that responds to commands
from the broker. If this task is dead, the device is not responding,
so we may as well cancel everything else.

An instance attribute `mainloop` contains this task, so that users
can see it, cancel it, or even add callbacks for the end of life.