don't crash with getHistory of 0 data
in the Good Old Times, so before 1.3.7-3-g6a2ffef (SVN 17899),
getting 0 data records with slotGetPropertyHistory meant to simply
get all the data. Since that, unfortunately, it means: crash the
server. This patch simply changes the default, so user's get a
response if they don't explicitly limit the number of returned records.