move the device info out of the device
each device has an info hash that summarizes some important details
about a device. It is so important that it is even needed after
the end of life of the device (to tell everyone who's gone). This
is why it is useful to put it outside of the device itself.