Take care that result of DataLogReader::navigateMetaRange is consistent.
I.e. non-existing index files in the final file range are accounted for with
zero records.
Also do not assign toFileNumber if not yet sure it exists...
We now can also fully trust the input file range and loop only up to last
given file number and not up to last existing.