add a converter to KaraboValues
Now every descriptor contains a method to convert data to
KaraboValues. This is very similar to the casting which we
did before. Just that before we were casting to numpy or
python builtin types, and now we are converting to our new
KaraboValues.

The old machinery is still there, as the GUI is based on it.
That may eventually go.