fix connectDevice in coroutines
connectDevice actually didn't work properly anywhere, but especially
so in coroutines. This fixes it and adds tests for it.

connectDevice was originally only designed for the command line, as
a simple and dirty replacement of getDevice. This changed, as we
realized that saving connections is actually a big burden with not
much reward.

Now connectDevice at least works, but it creates reference cycles.
I was not fast enough to fix those, so we need the garbage collector
to collect the devices.