select pasted elements
When a user pastes elements into the scene, the pasted elements should
be selected.

This patch is more complicated than it should be, as I realized that
Qt does not update layouts as it should. Forcing an update ("activate"
they call that a bit counter-intuitively in Qt) ourselves hopefully
also solves some other rendering problem.