add a two-motor example
this shows the use of the new DeviceNode. Note that I don't forsee any
necessity to connect the device node: this is done automatically for the
user.

Note also, that internally, we can access all properties of the device
node. The list of copied parameters is only for an outsider.