assure onDestruction is called by destructor
if a device is created and running and later is garbage collected,
we now still call onDestruction. This should only be possible
on the command line, as a device server always calls slotKillDevice
to destroy a device.

We take care that from the moment on that we don't exist anymore, we
cannot be called anymore, yet we can still receive responses to messages
sent to us.