Add a GitLab CI configuration to build dependencies
This also makes some changes to the dependency build script related to
building on the CI host and later downloading dependencies from an FTP server

The changes to the external dependency build script should not change the
behavior which is currently seen. Attempts to download the external deps only
occur with a clean checkout. Failure to do any of the fancy stuff should still
result in the current build behavior (ie: manually build everything)