Correct hack to accept empty VECTOR_STRING as VECTOR_HASH.
Both Python APIs serialise an empty list as VECTOR_STRING if not
told somehow that it should be an empty VECTOR_STRING.
So the validator has to accept this special case.
Previous validator code did also accept, but forgot to correct the type.
As a result, a VECTOR_HASH parameter of a device could in fact be a
VECTOR_STRING after device instantiation, making
Device::get<vector<Hash> >(..) fail with an exception.