remove StringList
as an empty vector of hashes and an empty vectors of strings are just an
empty list in Python, we had to mark empty lists. We used to mark vector
of strings, and take vector of hashes as the default.

This changed at some point, and vector of strings became the default.
Therefore we don't need to mark vectors of strings as such.