Improved auto_build_all.sh script
The script has a new option which allows to specify the number
of jobs for a make run. If not specified the max number of cores on
the system are used. Most importantly does this option also apply
during building the c++ tests. Those were previously build only using
a single job!!