fix race condition
This reimplements the core of our synchronization routines.

Offloading tasks into threads is done via our run_coroutine_or_thread,
which, for threads, uses asyncio's run_in_executor. The latter has no
notion of cancelling, as threads cannot easily be cancelled in a generic
way. We allow cancelling, as long as the task is stuck in a Karabo
routine. In this case we raise an exception in the task, which the user
code may handle. Finally, it will be handled by our code, which forwards
the exception to the creator of the task.