Bump the scene file version number to 2
The big change here is that icon data is directly written to the scene file
rather than stored in the project apart from the scene.

Both BaseIconsModel and DisplayIconset data models are affected. There are
still version 1 readers for both classes, but after reading data should be
upgraded before writing because it will be written as version 2.