This fix a subtile bug related to short-cut messaging
If a signal had more than one slot, it was always using the broker or p2p.
Even if one, more or all slots were directly callable did this happen.
While messages were still delivered the order of them was shuffled as
other communication patters (such as request/reponse) would do a faster
shortcutting and arrive earlier.

Now all slots of a signal are inspected for possible in-process shortcutting.