Fixes a serious bug happening during shortcut messaging.
In shortcut mode, the current code posted a message request directly to
the event-loop. Only after, a listener (sync of async) was registered in
order to receive the just sent reply. Sometimes (most likely depending on
the mood of the thread scheduler) a message was received before even starting
to wait for it. This resulted in sporadic test-failures.

This fix ensures having a listener registered before even sending the message and hence solves the issue.