SignalSlotable now uses a special message handler for p2p.
This fixes a performance bottleneck if many instances in  the same process
each consume many messages: tryToCallDirectly has to acquire a lock on a
static mutex - which is not needed since we can directly bind the correct
instance here.
That also means that karabo::net::ConsumeHandler does not need the first
string argument.

On the fly remove a useless string::substr(..) and fix a warning (in release
mode) concerning initialisation order in SignalSlotable constructor.