don't swallow debug printouts
if a printing to a remote console didn't work properly, the message was
swallowed sometimes. Let's at least print it to stdout, our last hope
to get it seen!

This is especially important as print is often used for debugging, and
so things might actually go wrong in an unexpected way.