Less restrictive locking of PointToPoint::Producer::m_registeredChannelsMutex
This is a "write seldom, read pretty, pretty often" map.
In fact, each SignalSlotable::doSendMessage locks this quasi static mutex
(there is only one PointToPoint instance per process).
So we protect reading by a 'read only' lock and use write-protection only
when needed.
The key of the map can well be an unordered_set.
An unordered_map is appropriate here as well, but causes compilation problems
that likely require to move the class declaration in (a separate?) header.

This is comparable to ec26fcaba56fa30bde761df9f56509b34aa6c55e which
addresses a similar problem on the broker part.