don't request configuration after schema update
after we received a new schema for a device proxy, we requested a new
configuration before actually applying the new schema, as we consider
the configuration outdated.

This has already been fixed on the sender side, as the sender guarantees
that a new proper configuration is sent immediately after a schema
change. This means that our code is not only superfluous, but even
dangerous: configuration changes will be received that refer to the new
schema, before we have changed.