avoid sending unnecessary changes
whenever someone sent a request to change a parameter on a middle
layer device, we shipped out the entire configuration. This is
completely exaggerated. Then normal procedure to ship out changes is
sufficient.

What we want instead is to assure all changes are shipped out at the
end of slotReconfigure. But this is easily done with a simple update.