reorganize the parents of a Configurable
until now, the parents of a Configurable had to be managed by the user.
This is an unnecessary burden on the user.

Instead, now we take care of the parent ourselves. And as we are already
at it, we actually make it possible that a Configurable has more than
one parent. Not that there was any urgent need for it, but it just falls
out easily.