Make config.orig zsh compatible.
On zsh, activating Karabo throwed an error:
.../karabo/var/config/config:70: START: assignment to invalid subscript range
since zsh arrays start indexing with 1.
To allow skipping index 0, karabo-check has to to a proper check:
In bash $START is the same as ${START[0]} if START is an array...
Now we check for the last array entry. If START does not exist or is a simple
variable, we karabo-check fails as well (with an additional other message).