Make connection to OutputChannel asynchronous since it blocks device start.
The remaining synchronous part was only the request to get the outputchannel
information Hash.
With the possibility to handle timeouts for requestAsync, this now can be
repeatedly tried if in the asynchronous pattern.
Since we since recently ensure that instanceNew is not called before channels
are there, a successful request that gives a negative answer does not have to
be foreseen anymore.