Provide possibility to register an error notifier, never just throw.
An error notifier can be given to readAsync which is called before the
message handler if something is fishy.
In addition, take care to never stop reading by just throwing: Since we post
to the event loop, the only effect of throwing is that the event loop thread
is kind of restarted after printing an error. That could cause an openmqc
zombie.
So now the JmsConsumer keeps trying to read until success and then calls the
message handler (that usually calls readAsync again).
The messagae is handler