Fix slow reading from broker by avoiding strands for blocking calls.
If a boost::asio::io_service::strand 'collides' with a strand used here for
the potentially blocking call to asyncConsumeMessage (since
MQReceiveMessageWithTimeout is blocking if no messages arrive), this means
that handlers posted to the strand can only be called when the blocking call
has finished.
Since the timeout in MQReceiveMessageWithTimeout is 100 ms, a JmsConsumer,
whose m_mqStrand collided with the one of a JmsConsumer that does not receive
any messages, receives at 10 Hz maximum.
See http://stackoverflow.com/questions/40291265/boostasio-reasoning-behind-num-implementations-for-io-servicestrand?rq=1
about 'colliding' strands.