Allow scenes to have widgets for injected properties
XFEL-CLOSED #17415

If a device is not in a project _and_ has injected properties, when opening
a scene which observes those injected properties, they would be missing widgets.
Now the device with the missing property will be monitored, and when schema
updates arrive, they will be checked for the missing properties. Once all
missing properties are filled in, the widget will pop into the scene.

Magic. Perhaps too magic.