Use asyncReply in slotStartDevice
1)
Avoids blocking work within the slot call: instantiating takes at least 1 s
which is the time we wait to get an answer if we must not start.
2)
Since the same slot cannot be executed in parallel, but block at their
slot mutex, 100 instantiateNoWait lead to thread starvation on the server.
3)
Synchronous instantiate works as before and gets its reply when done