make auto-reconnect work for pipelines
this is a very weird use case, but it made a bug reveal itself which
I wouldn't have found otherwise.

The start_channel coroutine in the NetworkInput removes itself from the
running task when it finishes, but gets removed by the code which
started it a second time, which is bound to fail.