Add more tests
Add tests on locks.
This lead to the change on the locking mechanism, and I dare hope
for the last time.

Two of the tests, `test_lock_not_owned` and `test_multiple_locks`
do not pass.
Whilst the behaviour is as intended in practice, these tests do not
reflect that.
Is it due to the cheap mock devices? Investigating will tell...