Update Online data test, add Offline data tests
Now, you're asking yourself, why such bad-taste naming?
Well, I'll tell you why:
My computer hung a number of times today, to the point that I had
to reboot it six times.
Why does it has to take two seconds for the `ls` command to work.
Is this 1978 or something?
Why does it has to take ten seconds to scroll a page in a pdf?
Is this 1978 or something?
Why does it has to take thirty seconds to display a menu in karabo?
Is this 1978 or something?
Why does it has to take thirty seconds between opening a terminal
and have a prompt?
Is this 1978 or something?

Can I get a real computer I can compile on, and not this `Kinder-
Surprise` joke?

I'll edit the names once I am in a better mood.

Can I get an SSD in the meanwhile?