Enable UserMacros to return an AcquiredData object
As to return an object from the coroutine that's added to the
eventloop, we give it a future, and return it.
This future is returned straightaway, and once the task goes from
`pending` to `finished`, the AcquiredData object can be accessed
using the `future.result()` function. Try it out so:

    >>> from karabo.usermacros import *
    >>> sc = AScan(Movable('m1','m2'),
                   [(1.0, 2.0), (2.0,3.0)],
                   'simcam', 0.1)
    >>> fut = sc.start()
    <Task pending coro=[...]>>
    >>> # once the scan is over:
    >>> fut
    <Task finished coro=[...]result=AcquiredOffline[...])>
    >>> data = fut.result()
    >>> data
    AcquiredOffline([...])