Robust logger instantiation if logger server starts in busy system.
Do not use blocking call to DeviceClient::exists(..), but rely on tracked
topology: If a data logger server starts into a running system, usually many
loggers need to be instantiated. They all send instanceNew, but that slot is
blocked by it being executed for the new server, so the other calls block
threads and the answer for slotPing (used inside 'exists(..)') cannot be
processed before the timeout inside 'exists', so 'exists' gives a fake
negative answer.