don't overwrite status in daemontools
if a service directory is locked, e.g. because a supervise is already
running, the new supervise, which tries to acquire the lock but fails
will still delete the status file of the other one. This fixes that.

we just stumbled over a bug in daemontools-encore! Instead of using this
only 7 years old, still buggy project we should have stayed with the
proven original daemontools, which had its last release 16 years
ago and doesn't contain this bug.