MDL: Test for NoneValue proxy in toSchemaAndAttrs
In middlelayer correctly check for "none-ness" of a proxy for a node in
toSchemaAndAttrs by using isSet to test whether it is a NoneValue as
well as testing for None.

Fixes issue when toSchemaAndAttrs is called for a device with
DeviceNodes where properties or slots are published to outside users.