Add PropertyProxy features needed by the scene view
This will also later be used by the configurator. Basically, we need
to have access to the value on the remote device, separate from the
local value stored in the binding. Additionally, we need to compare
the remote and local value to know if the local value should be sent
to the remote device.