Use individual thread for reading from broker to be safe against dead locks
While individual threads are not nice, here we protect the full Karabo
ecosystem (i.e. topic) by using one:
If every read is individually posted on the event loop and if a deadlock
anywhere (!) else in the code blocks all threads of the common event loop,
reading of broker messages would stop. Further messages are still received,
but not acknowledged and thus pile up on the broker: a black hole has begun...