Ensure that OutputChannel's TcpConnection is started.
This is a HACK and not even 100% failure proof.
What would be safe is to remove posting initializeServerConnection from
the constructor, but require that this is called (as a kind of second ctr.)
by 'the user' who creates the channel.
Maybe there is also a way to make the factory used in Configurator::create
ensure that there is a shared_ptr from the beginning?