GUI: Device Scene Links
This addresses the lack of a system to link the scenes provided
by the devices.
A scene link is currently only implemented between s without loading them into a project.

The features implemented will allow to:
- Have a Device scene link to its own other device-provided scenes.
- Middle Layer linking to the scenes it is linking to.
- Avoiding "schema evolution" in the project scenes.
  (e.g. a device provided scene evolves but its old remnant are living
  on in the project Database)

A displayType will be added to the VECTOR_STRING_ELEMENT
to facilitate scene building.