Longer queue thresholds in openmqc consumer, incl. clean-up. [ci-skip]
Current threshold is member variable to ensure that we continue to drop messages
to reach the lower threshold also after a non-droppable message was reached while
queue size was in between lower and higher threshold.

Avoid to call receiveQueue->size() several times since it enters a monitor
which I guess is something like a mutex lock.

Also generate log message with detailed info.