DeviceClient: always postpone device disconnection by 15 s
Before:
A device was always connected for at least 15 seconds. But after
this time, disconnecting from a device was done immediately when
the last client lost interest in it.

Now:
Once a GUI client connected to a device, connection will stay alive
until 15 seconds after the the client lost interest in the device.

Advantage:
Closing and quickly opening again scenes (when 'scrolling' through
scenes) should be smooth.

If a device was