do show the traceback after 10 s only
Show a traceback on where we got stuck if the event loops for more than
10 s. The event loop calls _run_once all the time, this is where we
check that.

OK, the problem was harder than initially thought. Asyncio actually
sleeps in _run_once. But we know the only place we are allowed to sleep:
in the selector.select. Just make an exception for that.