GUI: Sanitize macro scene widgets
On project "Save As" Macros, get a new UUID, if there is a scene
widget derived from them, they should be relinked.

This MR aims to solve this issue.

Also: Stopping user to "Save As" with the same name.
It is not completely foolproof, a nasty user can "save as"
and rename.