Gui server handles pipeline data via DeviceClient
In the long run this allows to make use of the automatic reconnect feature.

If we are lucky, this also solves the rare situations where a gui server
suddenly 'looses' big chunks of memory: The shared pointer to Gui client
connection was held twice befor this restructuring, so depending on
concurrency, such a pointer might have slipped removal and the connection
could survive - including big chunks of data from the async cache for big
detector data.