Make tool actions exclusive
The tool actions in the designer mode  are not as intuitive. When clicked, there's no indication that it is in 'Add <drawing>' mode. This commit adds exclusivity to these actions, such that one is selected, the button is checked and the cursor in the view is changed to cross. When the action is accepted, say successfully added a scene view, the toggle changes back to the selection tool (arrow pointer).