Fixed the testPipeMinData and testNPots test cases.
Both had an undocummented dependeny on the delay of the sender being 100 milliseconds.
The fix for not having such a dependency, which is the right solution, will be done in
a later merge request and will, most probably, involve fixes to the OutputChannel and
InputChannel classes. In particular the behavior of InputChannel regarding swapping of
the pod buffers and the handling of EndOfStream on both channel classes are pretty
unstable and subject to timing and racing conditions.