Rearranged the mutex structure - now working with two "interfering" mutexes.
The principle is that each activity is covered by its own mutex and when it has to do
something that interferes with the other activity it locks on that other activity mutex.
Telling from the current test results the rearrangement was effective - tests functionally
correct and total testing times decreased.