Fix for testMinData, code cleanup, lock for thread-safe access to m_isEndOfStream.
The testMinData failure when the sender has no delay was not related to race
conditions caused by wrong use of mutexes. Not testing whether the minData
condition had been reached before triggering the pot swap and the start of
processing for the new active pot at the end of InputChannel::triggerIOEvent
was the real reason for the test failure.

As part of the cleanup, some WARNs added on the previous commit remained as
TRACEs and the ones related to trace mutex acquisitions and releases have
been removed. InputChannel::swapBuffers has been removed - it consisted only
of a std::swap call. InputChannel::canCompute couldn't be removed because
it is public and referenced from the Karathon project.