Protect slot execution from parallel registration of more methods
Under normal circumstances this should not be a problem: A Slot is
usually registered in the constructor when messaging has not yet
started.
But request(..).receiveAsync creates temporary slots - there name is a
uuid. If that is not unique (fixed in MR !3302), it can happen that we
append a new method to an existing temporary slot while it is called
(to receive the reply for the first registered method...). So we can end
up looping on the container of functions in one thread and extending it
in another.
I guess that is what happens most of the times in the GUI server
crashes we observed:
- almost all crashes happened when big projects are opened, i.e. a lot
  of async requests to the project DB manager
- the two gdb back traces we have point to reply handling,
Note that a temporary slot that replies bool is used when setting up a
signal/slot connection to monitor a device.



See merge request Karabo/Framework!3307

(cherry picked from commit 7ca68c519c07c1ed965923b07d21d6a86c8c0e75)

d375e1a1 Protect slot execution from parallel registration of more methods