Fix for DeviceClient::getCurrentlyExectableCommands ignoring allowedStates.
The SlotElementBase class appended 'ConnectedSignals' as a child for every SlotElement. That caused the returned paths for slot elements in Schema::subSchemaByRules to have the form '[slot_name].conectedSignals'. Those paths with the connectedSignals suffix never had any 'allowedStates'  attributes linked to them in the schema and were always inserted into the resulting subschema, despite of the device's current state.