Avoid cyclic shared pointers in OutputChannel
Handler given to (Tcp)Channel and stored inside it must not bind-in a shared
pointer to the Channel itself because then the channel cannot be destroyed.
I think this happens when a connection is closed:
onTcpChannelError and thus onInputGone is called, but the Channel may still
keep a copy of the handler and thus a shared_ptr to itself.