FIX: ImageData setROIOffset should not test size
The ImageData cannot check that the ROI offset is within the
original image size, since it receives an already cut image.

For example: an image from a sensor 10k x 10k can have an ROI of (20, 20)
and the ImageData Object assumed that an offset bigger than 20 is wrong.
This fixes this bug.

XFEL-SOLVED#44330