CI: Project DB as a Docker image
Karabo does not `per se` run any java code, but relies on services
running on a java virtual machine.
This makes it difficult to run on platform with different java versions.

This removes the dependency on the existDB, and adapts the CI to run the
image that is provided here `https://git.xfel.eu/gitlab/ITDM/docker_existdb`
as a service. Which is runnign in the XFEL production system with a different
admin password.