Remove Darwin from build system
Karabo was never supported for the Darwin architecture.
In case it will ever be implemented, it will not be done by following the
current deployment system.
This removes the reference to this Platform from the karabo build process.
The references are still present for the dependency scripts.

Given the cleaning mood, I removed the serverControl directory altogether
it was a botched plan now substituted by the ansible deployment system