FIX: adapt to the latest Pint behaciour
Pint has changed its default behaviour in pretty printing.

Given the fact that the behavior change will only change some
visual results in html building and at best expose to the user
the fact that a QuantityValue built from a karabo `number` is
not a pure python number but a `dimensionless` Pint entity,
a change in the reference data in the tests has been implemented.