FEAT: a webserver to aggregate webservers
in a way that is completely independent from the control system
a webaggregatorserver can collect information from the webservers
in an installation and display it. This display can guide a
user to the correct web interface to restart a specific server.

In order to implement a completely asynchronous communication
the webservers are now making explicit use of the `asyncio` eventloop.