C++: Gui server schema sent to requesting channels
GUI server device sends the schema updates to all connected clients.
This allows the GUI client to get a schema despite the fact
that sometimes the DeviceClient of the GUI might not have a cached
schema to reply. The side effect of this is that during a large
device instantiation event all clients are affected.

This fixes that. A map of channels versus pending schemas is created
in a similar fashion as the `m_channels` map.
The GUI server bookkeeps the pending schemas and notifies only the
interested channels.