MDL: set server hostname
There can be technical reasons why a hostname cannot be set on the
target machines.

The hostname is taken from the host, and this is not guaranteed to be
resolved with a DNS service.
In addition, it can be beneficial to have the server appearing in the
topology under its secondary alias.

This addresses this issue in the Middlelayer API