BOUND: optionally set server hostname
The hostname is taken from the host, and this is not guaranteed to be resolvable as a host.
In addition, it can be beneficial to have the server appearing in the topology
under one of its secondary aliases.

This addresses this issue in the Bound API by adding an expected parameter to the
`DeviceServer`.

The name of this parameter is `hostName` which is in line with the parameter name
in the `Device` class.

[ci-skip]