Fix templated version if Hash::Node::setValue<Hash>
For setValue<Hash>(Hash(..)) we should not set the __classId either,
as for any other non-Hash-derived classes.

Best would be to get rid of the overload and specialise setValue<Hash> to be
as now the overlaoded, non-template version.

Schema::isCustomNode was previously based on the bug fixed here.

An effect was on updating the output schema of an OutputChannel after the
update of Device::updateSchema and Device::appendSchema for 2.7.0.
Now tested!