BUILD: Enable nightly build
The nightly build is a job run on schedules.
The job of the CI tasks is to check that the scheduled job is the one triggered.
Multiple Scheduled jobs can be implemented this way, e.g long connection tests,
coverage measurements.

The setting in the CI configurator is a variable named `SCHEDULED_JOB`
with a value describing the type of test.