CI: Conda cannot clean itself on Windows
The origin of the issue seems to be the fact that the main `dll` files are linked from the base environment.
For some reason, windows will lock the files in the environment to the executable of the CI (python) preventing the CI
from cleaning itself.

Here we choose the nuclear option of wiping everything clean.