OutputChannel callbacks with weak pointers
The call backs registered at output channels of devices are called as
long as the OutputChannel exists. These get deleted when the SignalSlotable
is destructed. So it could happen that the call back is called in one
thread and this->set locks a mutex of the Device. Then in another thread
the device gets deleted: Its destructor is called and then the destructors
of its members - among them the mutex locked in the first thread.
That results in an assertion:
.../extern/include/boost/thread/pthread/mutex.hpp:111: boost::mutex::~mutex(): Assertion `!res' failed.
The solution is to use bind_weak - or an equivalent treatment as chosen
here to avoid the need of a member function.