Refactor endOfStream treatment in C++ Input-/OutputChannel
Mark chunks in Memory to be endOfStream - but use this differently
in Input-/OutputChannel. The communication with each other is unchanged,
i.e. just a header sent via Tcp (to stay compatible with middlelayer
and old versions).

OutputChannel:
Mark a chunk as endOfStream and thus not containing data.
This is sent out immediately or put in the queue as normal chunks
from where it will be taken whenever the connected input is ready.
So signalEndOfStream is not blocking anymore (beyond Tcp::write).

InputChannel:
On endOfStream arrival, the inactive chunk is marked, though it can
still contain data. The endOfStream handler is called after that data
is processed.
Changed:
When endOfStream arrives, all remaining data is processed,
even if it does not respect minData requirement.
Known issue:
When endOfStream is received, but its processing could not yet
start, data that arrives now after the endOfStream may be processed
before the endOfStream.

Hijack to
- more robust test, i.e. increase likelihood of (expected) data loss
- improve formatting