C++: Input channels of devices show their connected output channels
Input channel nodes of C++ devices now have a property `"missingConnections"` that
tracks which of the "configured to be connected" output channels are currently not connected.
For bound Python, the property is (due to bindings) already there as well, but not yet filled.

Also re-organise the template specialisations for cppunit test extensions.