MDL: chatty plugin loading does not fail
As presented in https://in.xfel.eu/redmine/issues/74211, a module
loading that prints to stdout can ruin the schema parsing that
is performed through an external call by a middlelayer server.

This could potentially ruin an otherwise fine server.
This MR addresses this issue by redirecting the stdout on
plugin loading in the launcher.

Further more the plugins parsing that do not succeed are banned if the
external process has a return code other than 0.