TOOLS: activate actions can launch custom scripts
some environments might need extra care.
For example, the CentOS7 environment always requires the sourcing of
an enabling script to activate some version of the gcc compiler.
This is a bother for development on some systems.

This MR addresses this problem by allowing the user to launch a custom
scripts.

Several possibile uses of this include:
- allowing a developer to switch more easily between multiple versions of Karabo.
  On ``deactivate``, the servers could be shut down.
- make sure all plugins are installed on ``activate``
- make sure a set of servers are installed on ``activate``